Changed structure to Naomi plugin structure
It appears that this was originally written as a plugin for
the original Jasper. I have re-written the basic structure of
this as a jasper-dev plugin.

I am still working on the authentication/authorization stuff.
I have a separate example working which does list my current
calendar appointments. I should be able to adapt this to this
library pretty quickly. Unfortunately, it currently only works
if you can launch a web browser from your desktop, and the
method for setting up access requires having control of a domain
server.